+++ test/test-lock-devices.sh
@@ -123,7 +123,7 @@ new "Configure hostname on openconfig1"
 expectpart "$($clixon_cli -1 -f $CFG -E $CFD -m configure set devices device openconfig1 config system config hostname c-change)" 0 ""
 
 new "Commit 1"
-expectpart "$($clixon_cli -1 -f $CFG -E $CFD -m configure commit 2>&1)" 0 "Device openconfig1 in state PUSH_LOCK:protocol lock-denied Operation failed, lock is already held" --not-- "OK"
+expectpart "$($clixon_cli -1 -f $CFG -E $CFD -m configure commit 2>&1)" 0 "Device openconfig1 in state PUSH-LOCK:protocol lock-denied Operation failed, lock is already held" --not-- "OK"
 
 kill ${PIDS[0]}                   # kill the while loop above to close STDIN on 1st
 wait